Duplicate code reduction 2
